Move `stripSource` util from tests to nbdime-web
